Update Test.cpp
Tests passed! (code does what it looks like it should do).

However, "what it looks like it should do" can be improved greatly: as seen in output below, sentences come with unchomped spaces, and some are incomplete (due to use of dots not only as periods but also in abbreviations, etc. . . .)


OUTPUT:
Results for "white shark and tuna":
***********************************

 Most fish are ectothermic ("cold-blooded"), allowing their body temperatures to
 vary as ambient temperatures change, though some of the large active swimmers l
ike white shark and tuna can hold a higher core temperature.

Results for "fish is" and "fish are":
*************************************

A fish is any member of a paraphyletic group of organisms that consist of all gi
ll-bearing aquatic craniate animals that lack limbs with digits.

) than of other fish such as ray-finned fish or sharks, so the last common ances
tor of all fish is also an ancestor to tetrapods.



A typical fish is ectothermic, has a streamlined body for rapid swimming, extrac
ts oxygen from water using gills or uses an accessory breathing organ to breathe
 atmospheric oxygen, has two sets of paired fins, usually one or two (rarely thr
ee) dorsal fins, an anal fin, and a tail fin, has jaws, has skin that is usually
 covered with scales, and lays eggs.

 The larval period in oviparous fish is relatively short (usually only several w
eeks), and larvae rapidly grow and change appearance and structure (a process te
rmed metamorphosis) to become juveniles.



In biology --- the term fish is most strictly used to describe any animal with a
 backbone that has gills throughout life and has limbs, if any, in the shape of
fins.

[77]

In fisheries --- the term fish is used as a collective term, and includes mollus
ks, crustaceans and any aquatic animal which is harvested.